src/main/java/com/anirudhology/leetcoding/tree/BalancedBinaryTree.java
package com.anirudhology.leetcoding.tree;

import com.anirudhology.leetcoding.utils.TreeNode;

public class BalancedBinaryTree {

    public boolean isBalanced(TreeNode root) {
        // Special case
        if (root == null) {
            return true;
        }
        // Get heights of left and right subtrees
        int leftHeight = findHeight(root.left);
        int rightHeight = findHeight(root.right);
        if (Math.abs(leftHeight - rightHeight) > 1) {
            return false;
        }
        return isBalanced(root.left) && isBalanced(root.right);
    }

    private int findHeight(TreeNode root) {
        if (root == null) {
            return 0;
        }
        return 1 + Math.max(findHeight(root.left), findHeight(root.right));
    }
}
